Changed type of w and h to long.
git-svn-id: http://code.imagej.net/svn/imagej/trunk@4678 be23795f-a60e-40d4-ba19-055798872e57

+++ core/plugins/debug/src/main/java/imagej/core/plugins/debug/InvokePluginTest.java
@@ -66,8 +66,8 @@ public class InvokePluginTest implements ImageJPlugin {
 		final boolean signed = false;
 		final boolean floating = false;
 		final String fillType = "Ramp";
-		final int width = 512;
-		final int height = 512;
+		final long width = 512;
+		final long height = 512;
 		final Future<Module> future =
 			pluginService.run("imagej.io.plugins.NewImage", name, bitDepth, signed,
 				floating, fillType, width, height);
